refactor(execution): the snapshot's tests in their own file · the root's identity read is marked
The captured world's file crossed the 1500-line cap with the registry
test: its tests move to `snapshot/tests.rs` (the prologue precedent).
The project root's canonical read in the runtime config is the root's
identity, read once at composition — the same path the CLI canonicalizes
to judge a resume — and carries the seam-bypass marker with that reason.
